New function print_quadratic() and new module Quadratic Solver
All the print logic used to format the quadratic equations has been put
in the print_quadratic function. The commands asking for user input has
been placed in the main program in a file called QuadraticSolver.py The
original QuadraticSolve no longer asks for user input but just stores
functions to solve for the roots of quadratic equations and functions to
display the quadratic equations. QuadraticSolve.py is a library that is
called from the new QuadraticSolver.py file.