Fixed the case when a = 0. Added  a y-intercept function.
The functions print_quadratic, print_quad_roots, and y_intercept in
QuadraticSolve handle the case when a = 0. If the user sets a equal to
zero, the program prints out "This is not a quadratic equation".

A newly added function is the y_intercept function. The y_intercept
function prints out where the parabola corresponding to the quadratic
equation intersects the y-axis (where x = 0).