Added get_axis_of_symmetry(),get_vertex(),print_min_or_max(),print_vertex() functions
The get_axis_of_symmetry() function returns the axis of symmetry which
is a component of the quadratic formula (it is also the x-coordinate of
the vertex of the parabola that corresponds to a particular quadratic).
get_vertex() makes use of get_axis_of_symmetry to produce the parabola
vertex.

Printing functions were also created to print the vertex, the minimum or
maximum, and also print where the minimum or maximum occur. If the a
value is positive, then the parabola is concave up (cups up) and so the
quadratic equation has a minimum value. If the a value is negative, then
the parabola is concave down (cups down) and has a maximum value.