Added print_first_derivative and print_second_derivative functions
Added a couple of functions to print out the derivatives of the
quadratic function.

One instance where this could be useful is if you are working with the a
function that models the displacement of an object and the position of
the object can be described by a quadratic. The first derivative would
model the velocity and the second derivative would model the
acceleration of the object.