Added functions to print the sum and product of roots
For a quadratic equation of the form

y = ax^2 + bx + c

there is the property that

sum of the roots = -b / a

product of the roots = c / a

Now a couple of functions have been added to print out the sum and
product of the roots that are solutions to a given quadratic equation
based on user input. The current version only displays when the
solutions to the equation are real (no solutions containing the letter
'j').